Enum and scopes defined, example publish button.
