+++ prisma/schema.prisma
@@ -14,7 +14,7 @@ type ProtocolReviews {
 
 type ProtocolReviewsMethodologic {
   comments ProtocolReviewsComments[]
-  reviewer String                                @db.ObjectId
+  reviewer String                    @db.ObjectId
   veredict String
 }
 
@@ -30,16 +30,15 @@ type ProtocolReviewsScientific {
 
 type ProtocolReviewsScientificExternal {
   comments ProtocolReviewsComments[]
-  reviewer String                                      @db.ObjectId
+  reviewer String                    @db.ObjectId
   veredict String
 }
 
 type ProtocolReviewsScientificInternal {
   comments ProtocolReviewsComments[]
-  reviewer String                                      @db.ObjectId
+  reviewer String                    @db.ObjectId
   veredict String
 }
-
 
 type ProtocolSections {
   bibliography   ProtocolSectionsBibliography
@@ -133,18 +132,36 @@ type ProtocolSectionsPublication {
 }
 
 model protocol {
-  id           String           @id @default(auto()) @map("_id") @db.ObjectId
-  createdAt    DateTime         @db.Date
-  investigator String           @db.ObjectId 
-  reviews      ProtocolReviews?
-  sections     ProtocolSections
+  id         String           @id @default(auto()) @map("_id") @db.ObjectId
+  createdAt  DateTime         @db.Date
+  researcher String           @db.ObjectId
+  state      State            @default(DRAFT)
+  reviews    ProtocolReviews?
+  sections   ProtocolSections
 }
 
 model user {
   id        String   @id @default(auto()) @map("_id") @db.ObjectId
-  email     String
+  email     String   @unique
   lastLogin DateTime @db.Date
   name      String
   password  String
-  role      String
+  role      Role     @default(RESEARCHER)
+}
+
+enum Role {
+  RESEARCHER
+  SECRETARY
+  METHODOLOGIST
+  EVALUATOR
+  ADMIN
+}
+
+enum State {
+  NOT_CREATED
+  DRAFT
+  METHOD
+  SCIENTIFIC
+  ACCEPTED
+  ONGOING
 }
