+++ src/modules/protocol/elements/form/Select.tsx
@@ -1,8 +1,8 @@
 import { PropsWithChildren, useEffect, useState } from 'react'
 import { Combobox } from '@headlessui/react'
-import { useProtocolContext } from '../../../utils/createContext'
 import { Check, Selector, X } from 'tabler-icons-react'
 import clsx from 'clsx'
+import { useProtocolContext } from '@utils/createContext'
 
 export default function Select({
     path,
