+++ src/types/next-auth.d.ts
@@ -1,3 +1,4 @@
+import { user } from '@prisma/client'
 import { RoleType } from '@utils/zod'
 import { type DefaultSession } from 'next-auth'
 
@@ -6,9 +7,6 @@ declare module 'next-auth' {
      * Returned by `useSession`, `getSession` and received as a prop on the `SessionProvider` React Context
      */
     interface Session {
-        user?: {
-            id: string
-            role: RoleType
-        } & DefaultSession['user']
+        user?: user & DefaultSession['user']
     }
 }
