+++ src/utils/scopes.ts
@@ -1,23 +1,22 @@
+import { getServerSession } from 'next-auth'
+import { authOptions } from 'pages/api/auth/[...nextauth]'
 import {
-    RoleSchema,
+    ROLE,
     RoleType,
-    StateSchema,
+    STATE,
     StateType,
-    ActionSchema,
+    ACTION,
     ActionType,
+    ACCESS,
+    AccessType,
 } from './zod'
 
 // This component is meant to export helper functionalities in a centralized matter when we come to roles or states.
 // Atomic SRP Components that operate with one of these actions, should be guarded by this functions.
-
-const ROLE = RoleSchema.Enum
-const STATE = StateSchema.Enum
-const ACTION = ActionSchema.Enum
-
 // Check if role its allowed
 const ROLE_SCOPE = {
     [ROLE.RESEARCHER]: [
-        ACTION.LIST,
+        ACCESS.PROTOCOLS,
         ACTION.PUBLISH,
         ACTION.CREATE,
         ACTION.EDIT,
@@ -27,7 +26,8 @@ const ROLE_SCOPE = {
     [ROLE.METHODOLOGIST]: [ACTION.VIEW, ACTION.COMMENT],
     [ROLE.EVALUATOR]: [ACTION.VIEW, ACTION.COMMENT],
     [ROLE.ADMIN]: [
-        ACTION.LIST,
+        ACCESS.PROTOCOLS,
+        ACCESS.USERS,
         ACTION.CREATE,
         ACTION.VIEW,
         ACTION.EDIT,
@@ -62,7 +62,7 @@ export function canExecute(
 }
 
 // * Use when multiple protocols are involved so you can't pin point state (eg. Access to table, or to columns)
-export function canAccess(action: ActionType, role: RoleType) {
-    if (ROLE_SCOPE[role].some((a) => a === action)) return true
+export function canAccess(access: AccessType, role: RoleType) {
+    if (ROLE_SCOPE[role].some((a) => a === access)) return true
     return false
 }
